카테고리 페이지 구현 및 md파일 수정 완료

+++ src/main/java/com/exam/literaryplanner/domain/Board.java
@@ -11,7 +11,7 @@ public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "b_idx")
-    private Long bIdx;
+    private Integer bIdx;
 
     @Column(name = "b_type", nullable = false, length = 10)
     private String bType; // NOTICE / FAQ
@@ -32,8 +32,8 @@ public class Board {
 
     public Board() {}
 
-    public Long getBIdx() { return bIdx; }
-    public void setBIdx(Long bIdx) { this.bIdx = bIdx; }
+    public Integer getBIdx() { return bIdx; }
+    public void setBIdx(Integer bIdx) { this.bIdx = bIdx; }
 
     public String getBType() { return bType; }
     public void setBType(String bType) { this.bType = bType; }
